fix(python3.13): bump pynput version to 1.8.1 (#75)
Python 3.13 introduced a new `Thread` member `_handle`, which clashed
with the `_handle` name introduced by pynput's mixins.

Pynput introduced a fix for this in moses-palmer/pynput@f4cd9c5:

https://github.com/moses-palmer/pynput/commit/f4cd9c59f6a06763dc76d789c9a141a826ef23b1

This was released in `pynput` 1.7.8, but `aw-watcher-afk` was still using
`pynput` 1.7.7. This leads to the following errors at startup when using
a python 3.13 environment:

```
2025-06-26 08:19:44 [ERROR]: Unhandled exception in listener callback  (pynput.mouse.Listener:233)
Traceback (most recent call last):
  File "/home/user/repo/activitywatch/venv/lib/python3.13/site-packages/pynput/_util/__init__.py", line 229, in inner
    return f(self, *args, **kwargs)
  File "/home/user/repo/activitywatch/venv/lib/python3.13/site-packages/pynput/_util/xorg.py", line 470, in _handler
    self._handle(self._display_stop, event)
    ~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^
TypeError: '_thread._ThreadHandle' object is not callable
2025-06-26 08:19:49 [ERROR]: Unhandled exception in listener callback  (pynput.keyboard.Listener:233)
Traceback (most recent call last):
  File "/home/user/repo/activitywatch/venv/lib/python3.13/site-packages/pynput/_util/__init__.py", line 229, in inner
    return f(self, *args, **kwargs)
  File "/home/user/repo/activitywatch/venv/lib/python3.13/site-packages/pynput/_util/xorg.py", line 470, in _handler
    self._handle(self._display_stop, event)
    ~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^
TypeError: '_thread._ThreadHandle' object is not callable
```

There's a lot of noise in this change, as I appear to be using a newer
version of `poetry` than was last used to generate the lockfile.

To reproduce (perhaps with less noise), the command I used:

```bash
poetry update pynput
```